Avoid generating chunks beyond 100%
The batch enqueueing was queueing up chunks but not checking how many to queue up before it would hit 100%. As a result, users can see at times when the progress bar shows over 100% due to lingering extra chunks that were queued up.

The fix here it a bit checking heavy. Might be a way to simplify it. But it does solve the issue at hand!